+++ docs/CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Introduce new field "labels" in Flaw API (OSIDB-3803)
+
 ### Changed
 - Use keywords from ps-constants in CVEorg collector (OSIDB-3694)
+
+### Fixed
+- Fix workflow validation with conditional requirements (OSIDB-3524)
 
 ## [4.6.1] - 2024-12-06
 ### Fixed